fixed problem with truncated checksums
The method digest() returns (for sha256) 32-byte binary data
of type string. If that string contains a null byte '\0', the
underlying database writing methods truncate the string at
that null byte leading to a partial write of the checksum in
the database. In a successive check operation this leads to
problems. The problem did not occur unter Win7 but under
Linux64.