change TraverseDatabase to preorder traversal
SlowDelete was an academic example and it was the only
reason for having postorder traversal in TraverseDatabase
(you can only delete after deleting all descendants of
a node). All other methods (Export, Print, Check, Update)
do not care about the traversal type.Preorder traversal
is prefered now because we will have a TraverseDir
method too for usage in Import and Print. For Import we
need preorder traversal and for both new PrintDB and
PrintDir we need the same kind of traversal. So we will
use preorder traversal for all and get rid of SlowDelete.