first attempt to solve issue #3
Now there is an option to run update with correcting
checksum errors in the database by replacing the entries
in the database with the actual ones. It is called --fix.

There was an option like this before contained in the
NodeDB.Update method ("docheck" flag) but without access
via the command line. But the usage of that was not very
good because it just replaced all database content with
the actual data just like a re-import.

The problem is not completely solved now. It works fine
by running --update first, taking care of all minor problems
(missing files, new files) and detecting real errors.
After reviewing/fixing those you can run --fix to clean
up the rest. But this takes its time, rerunning all
the checksum calculations multiple times. Options are an
interactive mode to let the user decide what to do (might
be anfavorable too, because it requires user interaction
mixed with long waiting periods) or storing the problems
separately somewhere and just fixing those.